+++ com.devsquad10.order/src/main/java/com/devsquad10/order/domain/repository/OrderQuerydslRepository.java
@@ -1,0 +1,10 @@
+package com.devsquad10.order.domain.repository;
+
+import org.springframework.data.domain.Page;
+
+import com.devsquad10.order.domain.model.Order;
+
+public interface OrderQuerydslRepository {
+
+	Page<Order> findAll(String q, String category, int page, int size, String sort, String order);
+}
