+++ com.devsquad10.order/src/main/java/com/devsquad10/order/domain/repository/OrderRepository.java
@@ -3,8 +3,6 @@ package com.devsquad10.order.domain.repository;
 import java.util.Optional;
 import java.util.UUID;
 
-import org.springframework.data.domain.Page;
-
 import com.devsquad10.order.domain.model.Order;
 
 public interface OrderRepository {
@@ -13,5 +11,4 @@ public interface OrderRepository {
 
 	Order save(Order order);
 
-	Page<Order> findAll(String q, String category, int page, int size, String sort, String order);
 }
